add sweetalert in vuex

+++ client/src/store/index.js
@@ -2,8 +2,10 @@ import Vue from 'vue'
 import Vuex from 'vuex'
 import axios from 'axios'
 import router from '../router'
+import VueSweetalert2 from 'vue-sweetalert2'
 
 Vue.use(Vuex)
+Vue.use(VueSweetalert2)
 
 export default new Vuex.Store({
   state: {
@@ -23,6 +25,7 @@ export default new Vuex.Store({
       })
         .then(({ data }) => {
           commit('setProducts', data.data)
+          console.log(data.data)
         })
         .catch(err => {
           console.log(err)
@@ -35,9 +38,23 @@ export default new Vuex.Store({
       })
         .then(({ data }) => {
           localStorage.setItem('token', data.token)
+          this._vm.$swal.fire({
+            icon: 'success',
+            title: 'Login Success!',
+            showConfirmButton: false,
+            timer: 1500
+          })
           router.push({ name: 'Home' })
         })
-        .catch(console.log)
+        .catch(err => {
+          this._vm.$swal.fire({
+            icon: 'error',
+            title: 'Oops...',
+            text: err.response.data.errors[0],
+            showConfirmButton: false,
+            timer: 1500
+          })
+        })
     },
     addProduct (context, payload) {
       const { name, image_url, price, stock } = payload
@@ -49,23 +66,64 @@ export default new Vuex.Store({
         }
       })
         .then(({ data }) => {
-          console.log(data)
+          this._vm.$swal.fire({
+            toast: true,
+            position: 'top-end',
+            icon: 'success',
+            title: 'Add New Product Success!',
+            showConfirmButton: false,
+            timer: 2000
+          })
+          context.dispatch('fetchProduct')
         })
         .catch(err => {
+          this._vm.$swal.fire({
+            icon: 'error',
+            title: 'Oops...',
+            text: err.response.data.errors[0],
+            showConfirmButton: false,
+            timer: 1500
+          })
           console.log(err)
         })
     },
     deleteProduct (context, id) {
-      axios.delete(`http://localhost:3000/product/${id}`, {
-        headers: {
-          token: localStorage.token
-        }
-      })
-        .then(data => {
-          console.log(data)
-        })
-        .catch(err => {
-          console.log(err)
+      this._vm.$swal({
+        title: 'Are you sure want to delete this product?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+      }).then((result) => {
+          if(result.isConfirmed) {
+            axios.delete(`http://localhost:3000/product/${id}`, {
+              headers: {
+                token: localStorage.token
+              }
+            })
+            .then(({data}) => {
+                this._vm.$swal.fire({
+                  toast: true,
+                  position: 'top-end',
+                  icon: 'success',
+                  title: 'Delete Product Success!',
+                  showConfirmButton: false,
+                  timer: 2000
+                })
+                context.dispatch('fetchProduct')
+            })
+            .catch(err => {
+                this._vm.$swal.fire({
+                  icon: 'error',
+                  title: 'Oops...',
+                  text: err.response.data.errors[0],
+                  showConfirmButton: false,
+                  timer: 1500
+                })  
+            })
+          }
         })
     },
     updateProduct (context, payload) {
@@ -78,14 +136,36 @@ export default new Vuex.Store({
         }
       })
       .then(data => {
-        console.log(data)
+        this._vm.$swal.fire({
+          toast: true,
+          position: 'top-end',
+          icon: 'success',
+          title: 'Success Edit Product!',
+          showConfirmButton: false,
+          timer: 2000
+        })
+        context.dispatch('fetchProduct')
       })
       .catch(err => {
-        console.log(err)
+        this._vm.$swal.fire({
+          icon: 'error',
+          title: 'Oops...',
+          text: err.response.data.errors[0],
+          showConfirmButton: false,
+          timer: 1500
+        })
       })
     },
     logout () {
-      localStorage.clear
+      localStorage.clear()
+      this._vm.$swal.fire({
+        toast: true,
+        position: 'top-end',
+        icon: 'info',
+        title: 'Good bye, see you later!',
+        showConfirmButton: false,
+        timer: 2000
+      })
       router.push({ name: 'Login' })
     }
   },
